Add test for writing to broken .wal file
